chore(gui): Improved visualization of port statistics (#3569)
This PR fixes the issue where port counts overlap with each other when
two or more operators are close to each other by moving the output
port's text to the top of the edge and the input port's text to the
bottom of the edge so that they are differentiated. It also removes the
unnecessary texts that were used to address the port counts (ex:
output-0).

Changed File:
- core/gui/src/app/workspace/service/joint-ui/joint-ui.service.ts

<img width="376" height="223" alt="image"
src="https://github.com/user-attachments/assets/293fc586-0afa-4a7d-9488-55daada17d20"
/>